Moved functions around.
